SentinelExplorer

Documentation for SentinelExplorer.

SentinelExplorer.BoundingBox

SentinelExplorer.Point

SentinelExplorer.download_scene

SentinelExplorer.get_access_token

SentinelExplorer.get_scene_id

SentinelExplorer.search

SentinelExplorer.BoundingBox — Type

BoundingBox(ul, lr)

Construct a bounding box defined by the corners ul and lr.

All coordinates should be provided in latitude and longitude.

Parameters

ul: The upper-left corner of the box as a Tuple{T,T} of latitude and longitude.

lr: The lower-right corner of the box as a Tuple{T,T} of latitude and longitude.

Example

bb = BoundingBox((52.1, -114.4), (51.9, -114.1))

SentinelExplorer.Point — Type

Point(lat, lon)

Construct a point located at the provided latitude and longitude.

Parameters

lat: The latitude of the point.

lon: The longitude of the point.

Example

p = Point(52.0, -114.25)

SentinelExplorer.download_scene — Method

download_scene(scene, access_token; dir=pwd(), unzip=false)

Download the requested Sentinel scene using the provided access token.

Parameters

scene: The name of the Sentinel scene to download.

access_token: A token returned by a previous call to get_access_token.

Keywords

dir: The destination directory of the downloaded scene (default = pwd()).

unzip: If true, unzips and deletes the downloaded zip file (default = false).

Example

julia> scene = "S2B_MSIL2A_20200804T183919_N0500_R070_T11UPT_20230321T050221";

julia> get_scene_id(scene)
"29f0eaaf-0b15-412b-9597-16c16d4d79c6"

SentinelExplorer.get_access_token — Method

get_access_token()
get_access_token(username, password)

Authenticate with your Copernicus Data Space credentials.

The username and password may be passed explicitly, or provided as a pair of environment variables.

In the case of the latter, get_access_token() expects your username and password to be provided as the environment variables "SENTINELEXPLORERUSER" and "SENTINELEXPLORERPASS".

Parameters

username: Your Copernicus Data Space username.

password: Your Copernicus Data Space password.

Returns

An access token for downloading data.

Example

token = get_access_token(ENV["SENTINEL_EXPLORER_USER"], ENV["SENTINEL_EXPLORER_PASS"])
token = get_access_token()  # Same as Above

SentinelExplorer.get_scene_id — Method

get_scene_id(scene)

Lookup the unique identifier for the provided scene.

Parameters

scene: The name of the Sentinel scene to lookup.

Returns

The unique identifier for downloading the provided scene.

Example

julia> scene = "S2B_MSIL2A_20200804T183919_N0500_R070_T11UPT_20230321T050221";

julia> get_scene_id(scene)
"29f0eaaf-0b15-412b-9597-16c16d4d79c6"

SentinelExplorer.search — Method

search(satellite; product=nothing, dates=nothing, tile=nothing, clouds=nothing, geom=nothing, max_results=100)

Search for satellite images matching the provided filters.

Parameters

satellite: One of "SENTINEL-1", "SENTINEL-2", or "SENTINEL-3".

Keywords

product: The product type to search for such as "L2A", "L1C", "GRD", etc.

dates: The date range for image acquisition. Should be a tuple of DateTime objects.

tile: Restrict results to a given tile. Only available for Sentinel-2.

clouds: The maximum allowable cloud cover as a percentage. Not available for Sentinel-1.

geom: A geometry specifying the region of interest. Can be a Point, BoundingBox, or any other GeoInterface compatible geometry.

max_results: The maximum number of results to return (default = 100).

Returns

A DataFrame with the columns :Name, :AcquisitionDate, :PublicationDate, :CloudCover, and :Id.

Example

julia> geom = GeoDataFrames.read("test/data/roi.geojson").geometry[1];

julia> dates = (DateTime(2020, 8, 4), DateTime(2020, 8, 5));

julia> search("SENTINEL-2",  geom=geom, dates=dates)
3×5 DataFrame
 Row │ Name                               AcquisitionDate           Pub ⋯
     │ String                             String                    Str ⋯
─────┼───────────────────────────────────────────────────────────────────
   1 │ S2B_MSIL2A_20200804T183919_N0500…  2020-08-04T18:39:19.024Z  202 ⋯
   2 │ S2B_MSIL1C_20200804T183919_N0209…  2020-08-04T18:39:19.024Z  202
   3 │ S2B_MSIL1C_20200804T183919_N0500…  2020-08-04T18:39:19.024Z  202
                                                        3 columns omitted